Add spec for admin anime body
